# api/gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/Raina-Hardik/smelt/internal/workflow"
	"github.com/oapi-codegen/runtime"
)

// Defines values for RunStatus.
const (
	RunStatusCancelled RunStatus = "cancelled"
	RunStatusDone      RunStatus = "done"
	RunStatusFailed    RunStatus = "failed"
	RunStatusRunning   RunStatus = "running"
)

// Valid indicates whether the value is a known member of the RunStatus enum.
func (e RunStatus) Valid() bool {
	switch e {
	case RunStatusCancelled:
		return true
	case RunStatusDone:
		return true
	case RunStatusFailed:
		return true
	case RunStatusRunning:
		return true
	default:
		return false
	}
}

// Defines values for RunDetailStatus.
const (
	RunDetailStatusCancelled RunDetailStatus = "cancelled"
	RunDetailStatusDone      RunDetailStatus = "done"
	RunDetailStatusFailed    RunDetailStatus = "failed"
	RunDetailStatusRunning   RunDetailStatus = "running"
)

// Valid indicates whether the value is a known member of the RunDetailStatus enum.
func (e RunDetailStatus) Valid() bool {
	switch e {
	case RunDetailStatusCancelled:
		return true
	case RunDetailStatusDone:
		return true
	case RunDetailStatusFailed:
		return true
	case RunDetailStatusRunning:
		return true
	default:
		return false
	}
}

// Defines values for ListRunsParamsStatus.
const (
	Done    ListRunsParamsStatus = "done"
	Failed  ListRunsParamsStatus = "failed"
	Running ListRunsParamsStatus = "running"
)

// Valid indicates whether the value is a known member of the ListRunsParamsStatus enum.
func (e ListRunsParamsStatus) Valid() bool {
	switch e {
	case Done:
		return true
	case Failed:
		return true
	case Running:
		return true
	default:
		return false
	}
}

// Action What to do with a file when all conditions match.
type Action = workflow.Action

// Cond A single per-file condition. Not every field/op combination is valid:
// string fields (codec, audio, ext) accept only eq/ne; numeric fields
// (height, width, bitrate, duration) accept the full comparison set.
type Cond = workflow.Cond

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// Health defines model for Health.
type Health struct {
	Status string `json:"status"`
}

// Job defines model for Job.
type Job struct {
	ID         int64     `json:"id"`
	Pct        float64   `json:"pct"`
	SourcePath string    `json:"source_path"`
	Speed      string    `json:"speed,omitempty"`
	Status     string    `json:"status"`
	UpdatedAt  time.Time `json:"updated_at"`
}

// Program defines model for Program.
type Program struct {
	CreatedAt time.Time `json:"created_at"`
	Ext       []string  `json:"ext"`
	ID        string    `json:"id"`
	Name      string    `json:"name"`
	Rules     []Rule    `json:"rules"`
	Schedule  string    `json:"schedule"`
	Src       string    `json:"src"`
	UpdatedAt time.Time `json:"updated_at"`
}

// ProgramInput defines model for ProgramInput.
type ProgramInput struct {
	// Ext File extensions to match. Defaults to mkv, mp4, avi when empty.
	Ext   *[]string `json:"ext,omitempty"`
	Name  *string   `json:"name,omitempty"`
	Rules *[]Rule   `json:"rules,omitempty"`

	// Schedule Optional cron expression.
	Schedule *string `json:"schedule,omitempty"`

	// Src Root directory the program scans.
	Src string `json:"src"`
}

// Rule Ordered decision rule — first rule whose conditions all match wins.
type Rule = workflow.Rule

// Run defines model for Run.
type Run struct {
	Failed     int        `json:"failed"`
	FinishedAt *time.Time `json:"finished_at,omitempty"`
	Name       string     `json:"name"`
	OK         int        `json:"ok"`
	RunID      string     `json:"run_id"`
	Skipped    int        `json:"skipped"`
	StartedAt  time.Time  `json:"started_at"`
	Status     RunStatus  `json:"status"`
	Total      int        `json:"total"`
}

// RunStatus defines model for Run.Status.
type RunStatus string

// RunAccepted defines model for RunAccepted.
type RunAccepted struct {
	RunID string `json:"run_id"`
}

// RunDetail defines model for RunDetail.
type RunDetail struct {
	Failed     int             `json:"failed"`
	FinishedAt *time.Time      `json:"finished_at,omitempty"`
	Jobs       []Job           `json:"jobs"`
	Name       string          `json:"name"`
	OK         int             `json:"ok"`
	RunID      string          `json:"run_id"`
	Skipped    int             `json:"skipped"`
	StartedAt  time.Time       `json:"started_at"`
	Status     RunDetailStatus `json:"status"`
	Total      int             `json:"total"`
}

// RunDetailStatus defines model for RunDetail.Status.
type RunDetailStatus string

// ProgramID defines model for ProgramID.
type ProgramID = string

// BadRequest defines model for BadRequest.
type BadRequest = Error

// InternalError defines model for InternalError.
type InternalError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// ListRunsParams defines parameters for ListRuns.
type ListRunsParams struct {
	Limit  *int                  `form:"limit,omitempty" json:"limit,omitempty"`
	Status *ListRunsParamsStatus `form:"status,omitempty" json:"status,omitempty"`
}

// ListRunsParamsStatus defines parameters for ListRuns.
type ListRunsParamsStatus string

// CreateProgramJSONRequestBody defines body for CreateProgram for application/json ContentType.
type CreateProgramJSONRequestBody = ProgramInput

// UpdateProgramJSONRequestBody defines body for UpdateProgram for application/json ContentType.
type UpdateProgramJSONRequestBody = ProgramInput

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Liveness check
	// (GET /api/health)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// List all stored programs
	// (GET /api/programs)
	ListPrograms(w http.ResponseWriter, r *http.Request)
	// Create a program
	// (POST /api/programs)
	CreateProgram(w http.ResponseWriter, r *http.Request)
	// Delete a program
	// (DELETE /api/programs/{id})
	DeleteProgram(w http.ResponseWriter, r *http.Request, id ProgramID)
	// Get a single program
	// (GET /api/programs/{id})
	GetProgram(w http.ResponseWriter, r *http.Request, id ProgramID)
	// Replace a program's fields
	// (PUT /api/programs/{id})
	UpdateProgram(w http.ResponseWriter, r *http.Request, id ProgramID)
	// Render and execute the program as a background run
	// (POST /api/programs/{id}/run)
	RunProgram(w http.ResponseWriter, r *http.Request, id ProgramID)
	// List recent runs
	// (GET /api/runs)
	ListRuns(w http.ResponseWriter, r *http.Request, params ListRunsParams)
	// Cancel a running subprocess (SIGTERM to its process group)
	// (DELETE /api/runs/{id})
	CancelRun(w http.ResponseWriter, r *http.Request, id string)
	// Get a run with its full job list
	// (GET /api/runs/{id})
	GetRun(w http.ResponseWriter, r *http.Request, id string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListPrograms operation middleware
func (siw *ServerInterfaceWrapper) ListPrograms(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListPrograms(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateProgram operation middleware
func (siw *ServerInterfaceWrapper) CreateProgram(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateProgram(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteProgram operation middleware
func (siw *ServerInterfaceWrapper) DeleteProgram(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ProgramID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteProgram(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetProgram operation middleware
func (siw *ServerInterfaceWrapper) GetProgram(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ProgramID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetProgram(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateProgram operation middleware
func (siw *ServerInterfaceWrapper) UpdateProgram(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ProgramID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateProgram(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RunProgram operation middleware
func (siw *ServerInterfaceWrapper) RunProgram(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ProgramID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RunProgram(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListRuns operation middleware
func (siw *ServerInterfaceWrapper) ListRuns(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ListRunsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "status", r.URL.Query(), &params.Status, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "status"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "status", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListRuns(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CancelRun operation middleware
func (siw *ServerInterfaceWrapper) CancelRun(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CancelRun(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRun operation middleware
func (siw *ServerInterfaceWrapper) GetRun(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRun(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/health", wrapper.GetHealth)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/programs", wrapper.ListPrograms)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/programs", wrapper.CreateProgram)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/api/programs/{id}", wrapper.DeleteProgram)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/programs/{id}", wrapper.GetProgram)
	m.HandleFunc(http.MethodPut+" "+options.BaseURL+"/api/programs/{id}", wrapper.UpdateProgram)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/programs/{id}/run", wrapper.RunProgram)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/runs", wrapper.ListRuns)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/api/runs/{id}", wrapper.CancelRun)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/runs/{id}", wrapper.GetRun)

	return m
}